Code review adjustments
