+++ app/src/main/java/org/thoughtcrime/securesms/util/ViewUtilities.kt
@@ -9,12 +9,15 @@ import android.graphics.Bitmap
 import android.graphics.PointF
 import android.graphics.Rect
 import android.util.Size
+import android.util.TypedValue
 import android.view.View
 import androidx.annotation.ColorInt
 import androidx.annotation.DimenRes
 import network.loki.messenger.R
 import org.session.libsession.utilities.getColorFromAttr
 import android.view.inputmethod.InputMethodManager
+import androidx.annotation.AttrRes
+import androidx.annotation.ColorRes
 import androidx.core.graphics.applyCanvas
 import kotlin.math.roundToInt
 
@@ -31,6 +34,20 @@ val View.hitRect: Rect
 
 @ColorInt
 fun Context.getAccentColor() = getColorFromAttr(R.attr.colorAccent)
+
+// Method to grab the appropriate attribute for a message colour.
+// Note: This is an attribute, NOT a resource Id - see `getColorResourceIdFromAttr` for that.
+@AttrRes
+fun getMessageTextColourAttr(messageIsOutgoing: Boolean): Int =
+    if (messageIsOutgoing) R.attr.message_sent_text_color else R.attr.message_received_text_color
+
+// Method to get an actual R.id.<SOME_COLOUR> resource Id from an attribute such as R.attr.message_sent_text_color etc.
+@ColorRes
+fun getColorResourceIdFromAttr(context: Context, attr: Int): Int {
+    val typedValue = TypedValue()
+    context.theme.resolveAttribute(attr, typedValue, true)
+    return typedValue.resourceId
+}
 
 fun View.animateSizeChange(@DimenRes startSizeID: Int, @DimenRes endSizeID: Int, animationDuration: Long = 250) {
     val startSize = resources.getDimension(startSizeID)
